# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CommunityLinkGroupRequest:
    properties:
      childJid:
        type: string
      parentJid:
        type: string
    required:
    - childJid
    - parentJid
    type: object
...
  dto.CommunityUpdateRequestParticipantsRequest:
    properties:
      action:
        enum:
        - approve
        - reject
        type: string
      communityJid:
        type: string
      participants:
        items:
          type: string
        type: array
    required:
    - action
    - communityJid
    - participants
    type: object
...
  dto.CreateCommunityRequest:
    properties:
      description:
        type: string
      subject:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - subject
    type: object
  dto.CreateCommunitySubGroupRequest:
    properties:
      parentJid:
        type: string
      participants:
        items:
          type: string
        type: array
      subject:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - parentJid
    - participants
    - subject
    type: object
...
  utils.AuthenticationErrorResponse:
    properties:
      message:
        example: Unauthorized
        type: string
    type: object
  utils.HTTPErrorResponse:
    properties:
      error:
        description: |-
          Error is a legacy opaque value. Go's JSON encoder does not serialize an
          error message here, so clients should use Message and ErrorMessage.
        type: object
      errorMessage:
        type: string
      message:
        type: string
    type: object
...
  whatsmiau.GroupInfoResponse:
    properties:
      announce:
        type: boolean
      creation:
        type: integer
      desc:
        type: string
      descId:
        type: string
      ephemeral:
        type: integer
      groupAddMode:
        type: string
      id:
        type: string
      isCommunity:
        type: boolean
      isCommunityAnnounce:
        type: boolean
      joinApprovalMode:
        type: boolean
      linkedParent:
        type: string
      memberAddMode:
        type: string
      owner:
        type: string
      participants:
        items:
          $ref: '#/definitions/whatsmiau.GroupParticipantResponse'
        type: array
      restrict:
        type: boolean
      size:
        type: integer
      subject:
        type: string
      subjectOwner:
        type: string
      subjectTime:
        type: integer
    type: object
...
  whatsmiau.LinkedParticipantResponse:
    properties:
      jid:
        type: string
      lid:
        type: string
    type: object
  whatsmiau.RequestParticipantResponse:
    properties:
      jid:
        type: string
      lid:
        type: string
      requestedAt:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    name: Verbeux AI
    url: https://github.com/verbeux-ai/whatsmiau
  description: Authenticated WhatsApp API compatible with Evolution API. The specification
    covers the complete application route surface, including canonical routes, compatibility
    aliases, calls and documentation delivery.
  license:
    name: MIT
  title: WhatsMiau API
  version: 0.3.2
paths:
...
  /v1/instance/{instance}/community/create:
    post:
      consumes:
      - application/json
      operationId: postV1InstanceByInstanceCommunityCreate
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Community payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommunityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/whatsmiau.GroupInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a WhatsApp community (parent group)
      tags:
      - Community
  /v1/instance/{instance}/community/createSubGroup:
    post:
      consumes:
      - application/json
      operationId: postV1InstanceByInstanceCommunityCreateSubGroup
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Sub group payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommunitySubGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/whatsmiau.GroupInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new group already linked to a community
      tags:
      - Community
  /v1/instance/{instance}/community/linkGroup:
    post:
      consumes:
      - application/json
      operationId: postV1InstanceByInstanceCommunityLinkGroup
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Link payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CommunityLinkGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Link an existing group to a community
      tags:
      - Community
  /v1/instance/{instance}/community/linkedGroupsParticipants:
    get:
      operationId: getV1InstanceByInstanceCommunityLinkedGroupsParticipants
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Community JID
        in: query
        name: communityJid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/whatsmiau.LinkedParticipantResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List participants across all linked groups of a community
      tags:
      - Community
  /v1/instance/{instance}/community/requestParticipants:
    get:
      operationId: getV1InstanceByInstanceCommunityRequestParticipants
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Community JID
        in: query
        name: communityJid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/whatsmiau.RequestParticipantResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List pending join requests
      tags:
      - Community
  /v1/instance/{instance}/community/requestParticipants/update:
    post:
      consumes:
      - application/json
      operationId: postV1InstanceByInstanceCommunityRequestParticipantsUpdate
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CommunityUpdateRequestParticipantsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/whatsmiau.UpdateParticipantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Approve or reject pending join requests
      tags:
      - Community
  /v1/instance/{instance}/community/setGroupAddMode:
    post:
      consumes:
      - application/json
      description: Use admin_add to restrict adding/linking groups to community admins,
        or all_member_add to allow any community member.
      operationId: postV1InstanceByInstanceCommunitySetGroupAddMode
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Community group-add mode payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CommunitySetAddModeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Set who can add groups to a community
      tags:
      - Community
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apikey
    type: apiKey
swagger: "2.0"